requirements.txt
Django==1.8.4
django-allauth==0.23.0
djangorestframework==3.2.3
oauthlib==1.0.3
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
resource
deployer